Tests/LibWeb/Text/expected/indexeddb-commit-prevents-new-requests.txt

put.success
second put threw: TransactionInactiveError
tx.complete
DONE
